--- packages/view3d/src/plugin/PostProcessing/index.ts
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2020 NAVER Corp.
- * egjs projects are licensed under the MIT license
- */
-import PostProcessing, { PostProcessingOptions } from "./PostProcessing";
-
-
-export {
-  PostProcessing
-};
-
-export type {
-  PostProcessingOptions
-};
